META-3040 Restrict mutations via entity/relationship APIs

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/AtlasRelationshipStoreV2.java
@@ -69,6 +69,7 @@ import java.util.Set;
 import java.util.UUID;
 
 import static org.apache.atlas.AtlasConfiguration.NOTIFICATION_RELATIONSHIPS_ENABLED;
+import static org.apache.atlas.accesscontrol.AccessControlUtil.ensureNonAccessControlRelType;
 import static org.apache.atlas.model.instance.AtlasEntity.Status.DELETED;
 import static org.apache.atlas.model.typedef.AtlasRelationshipDef.PropagateTags.BOTH;
 import static org.apache.atlas.model.typedef.AtlasRelationshipDef.PropagateTags.NONE;
@@ -137,6 +138,8 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
         if (LOG.isDebugEnabled()) {
             LOG.debug("==> update({})", relationship);
         }
+
+        ensureNonAccessControlRelType(relationship.getTypeName());
 
         String guid = relationship.getGuid();
 
@@ -309,6 +312,8 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
                 throw new AtlasBaseException(AtlasErrorCode.RELATIONSHIP_ALREADY_DELETED, guid);
             }
 
+            ensureNonAccessControlRelType(getTypeName(edge));
+
             edgesToDelete.add(edge);
             deletedRelationships.add(entityRetriever.mapEdgeToAtlasRelationship(edge));
         }
@@ -346,7 +351,6 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
 
         AtlasEdge edge = graphHelper.getEdgeForGUID(guid);
 
-
         if (edge == null) {
             throw new AtlasBaseException(AtlasErrorCode.RELATIONSHIP_GUID_NOT_FOUND, guid);
         }
@@ -354,6 +358,8 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
         if (getState(edge) == DELETED) {
             throw new AtlasBaseException(AtlasErrorCode.RELATIONSHIP_ALREADY_DELETED, guid);
         }
+
+        ensureNonAccessControlRelType(getTypeName(edge));
         deleteDelegate.getHandler().resetHasLineageOnInputOutputDelete(Collections.singleton(edge), null);
         deleteDelegate.getHandler().deleteRelationships(Collections.singleton(edge), forceDelete);
 
@@ -423,6 +429,8 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
 
     public AtlasEdge createRelationship(AtlasVertex end1Vertex, AtlasVertex end2Vertex, AtlasRelationship relationship, boolean existingRelationshipCheck) throws AtlasBaseException {
         AtlasEdge ret;
+
+        ensureNonAccessControlRelType(relationship.getTypeName());
 
         try {
             validateRelationship(end1Vertex, end2Vertex, relationship);
